- event.c:event_time_set(): fix compile warning
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/trunk@8911 82007160-df01-0410-b94d-b575c5fd34c7

+++ src/event.c
@@ -124,7 +124,7 @@ event_time_set(void)
 
   if (clock_gettime(CLOCK_REALTIME, &newtime))
     exit(EXIT_FAILURE);
-  else if (event_base->time.sec_real > newtime.tv_sec)
+  else if (event_base->time.sec_real > (uintmax_t)newtime.tv_sec)
     ilog(LOG_TYPE_IRCD, "System clock is running backwards - (%ju < %ju)",
          (uintmax_t)newtime.tv_sec, event_base->time.sec_real);
   else
